Pneumoniax 3000/requirements.txt
streamlit>=1.28.0
pydicom>=2.4.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
tensorflow>=2.11.0
keras>=2.11.0

